Move browser tests fixtures to a dedicated conftest.py file in their subdirectory
